Clean up deploy method
